afxdp/worker_runtime: relocate inline tests to worker_runtime_tests.rs
Same playbook. 55-LOC inline test block extracted. worker_runtime.rs: 225 → 173 LOC.
cargo build/test 92/878.
